requirements.txt
pymongo
openai
openai-agents
tiktoken
PyPDF2
python-docx
boto3
pydantic-settings
tenacity
